refactor(ui): extract coordinator-admin/components/join-requests-panel.ts
Move renderJoinRequestsPanel into its own module. The panel now takes a
deps object with { summary, reviewJoinRequest } so the barrel wires in
the factory action.
